--- api/src/main/resources/db/migration/V1.0.0__ADD_IS_FINAL_EXPERIMENT_TO_ITEM.sql
@@ -1,2 +1,0 @@
-alter table item
-	add final_experiment bit(1) default 0;
