+++ api/src/test/java/club/yunzhi/workhome/controller/WorkControllerTest.java
@@ -8,16 +8,30 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.core.io.Resource;
+import org.springframework.core.io.ResourceLoader;
 import org.springframework.http.MediaType;
+import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.web.servlet.MockMvcBuilder;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
+import org.springframework.util.ResourceUtils;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.multipart;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 class WorkControllerTest extends ControllerTest {
+    private static final Logger logger = LoggerFactory.getLogger(WorkControllerTest.class);
+
     private String url = "/work";
     @MockBean
     WorkService workService;
+    @Autowired
+    private ResourceLoader loader;
 
     @Test
     void getByItemId() throws Exception {
@@ -68,5 +82,28 @@ class WorkControllerTest extends ControllerTest {
         ArgumentCaptor<Work> workArgumentCaptor = ArgumentCaptor.forClass(Work.class);
         Mockito.verify(this.workService).updateOfCurrentStudent(Mockito.eq(workId), workArgumentCaptor.capture());
         Assertions.assertEquals(workId.compareTo(workArgumentCaptor.getValue().getId()), 0);
+    }
+
+
+    @Test
+    void uploadWork() throws Exception {
+        logger.debug("定义常量");
+        final String NAME = "attachment";
+        final String FILE_NAME = "example.jpeg";
+
+        logger.debug("获取资源");
+        Resource resource = loader.getResource(ResourceUtils.CLASSPATH_URL_PREFIX + FILE_NAME);
+
+        logger.debug("创建模拟文件");
+        MockMultipartFile multipartFile = new MockMultipartFile(NAME, FILE_NAME, "image/jpeg", resource.getInputStream());
+
+        logger.debug("断言请求成功");
+        this.mockMvc.perform(multipart(this.url + "/uploadWork")
+                .file(multipartFile)
+                .param("option1", "")
+        ).andExpect(status().isOk());
+
+        logger.debug("断言方法调用成功");
+        Mockito.verify(workService).uploadWork(multipartFile, "", null);
     }
 }
